ii-git: Change tabs to spaces, add prepare()
I'm going to be using this as an example PKGBUILD for CLUG, so it should
probably follow everything I say. There is also no use for the config.mk
source file, It's included in the tarball and edited in that dir also.
People that want to edit the config.mk can add it back in or put in a
.diff for that.